feat(DTFS2-7232): move amendments query types to libs/common
