--- libs/common/src/constants/amendment-status.ts
@@ -1,5 +1,0 @@
-export const AMENDMENT_STATUS = {
-  NOT_STARTED: 'Not started',
-  IN_PROGRESS: 'In progress',
-  COMPLETED: 'Completed',
-} as const;
