+++ libs/common/src/constants/index.ts
@@ -11,7 +11,7 @@ export * from './fee-record-status';
 export * from './audit-user-types';
 export * from './company-registration-number';
 export * from './keying-sheet-row-status';
-export * from './amendment-status';
+export * from './amendments';
 export * from './api-error-code';
 export * from './facility-type';
 export * from './facility-provided-details';
