+++ libs/common/src/types/amendment-status.ts
@@ -1,4 +1,4 @@
-import { AMENDMENT_STATUS } from '../constants/amendment-status';
+import { AMENDMENT_STATUS } from '../constants/amendments';
 import { ValuesOf } from './types-helper';
 
 export type AmendmentStatus = ValuesOf<typeof AMENDMENT_STATUS>;
